pyproject.toml
[project]
name = "remora-test"
version = "0.1.0"
description = "Demonstration repository for remora-v2 runtime capabilities."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
  { name = "Bullish Design", email = "BullishDesignEngineering@gmail.com" },
]

dependencies = [
  "pydantic>=2.12.5",
  "pyyaml>=6.0.2",
  "remora[all-languages,search-local,lsp]",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.1",
  "mypy>=1.10",
  "ruff>=0.5.0",
]

[tool.uv.sources]
#remora = { path = "../remora-v2", editable = false }
remora = { git = "https://github.com/Bullish-Design/remora-v2.git", tag="v0.7.7" } # rev = "cd93213efd6a99cf45a3c9e19fb34e66f4377450" }


[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
  "src",
  "README.md",
  "tests",
  "pyproject.toml",
]

[tool.uv]

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=60"
testpaths = ["tests"]
markers = [
  "live: tests that require a running remora runtime and optional external services",
]

[tool.mypy]
python_version = "3.13"
files = ["src"]
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
no_implicit_optional = true

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
